fix: client display of login rewards
While the server was correctly picking the right login rewards tier, based off the login streak of the player, the client was instead displaying the icon/ui for tier+1 (days 0-2 were showing as days 3-7, etc). This is due to the fact that the reward definitions have the tiers starting at 1, but the client expects that the tiers start at 0.

exising reward definition documentation has been cleaned up .